Fix file_lenght SwiftLint warning
Glia.swift has been separated and functions for starting
engagement have been moved to separate file to fix
SwiftLint warning about size of the class body.
