+++ GliaWidgets/Public/Glia/Glia.swift
@@ -1,5 +1,5 @@
-import UIKit
 import SalemoveSDK
+import UIKit
 
 /// Engagement media type.
 public enum EngagementKind: Equatable {
@@ -39,10 +39,8 @@ public protocol SceneProvider: AnyObject {
 public class Glia {
     /// A singleton to access the Glia's interface.
     public static let sharedInstance = Glia(environment: .live)
-
     /// Current engagement media type.
     public var engagement: EngagementKind { return rootCoordinator?.engagementKind ?? .none }
-
     /// Used to monitor engagement state changes.
     public var onEvent: ((GliaEvent) -> Void)?
 
@@ -62,7 +60,8 @@ public class Glia {
             screenShareHandler: environment.screenShareHandler,
             audioSession: environment.audioSession,
             date: environment.date,
-            engagedOperator: { [weak self] in self?.environment.coreSdk.getCurrentEngagement()?.engagedOperator
+            engagedOperator: { [weak self] in
+                self?.environment.coreSdk.getCurrentEngagement()?.engagedOperator
             }
         )
     )
@@ -98,10 +97,7 @@ public class Glia {
             with: sdkConfiguration,
             queueID: queueId,
             visitorContext: visitorContext,
-            environment: .init(
-                coreSdk: environment.coreSdk,
-                gcd: environment.gcd
-            )
+            environment: .init(coreSdk: environment.coreSdk, gcd: environment.gcd)
         )
 
         interactor = createdInteractor
@@ -120,18 +116,14 @@ public class Glia {
             guard let engagement = self?.environment.coreSdk.getCurrentEngagement(), engagement.source == .callVisualizer else {
                 return
             }
-
             switch event {
             case .screenShareOffer(answer: let answer):
                 self?.environment.coreSdk.requestEngagedOperator { operators, _ in
                     self?.callVisualizer.offerScreenShare(
                         from: operators ?? [],
                         configuration: Theme().alertConfiguration.screenShareOffer,
-                        accepted: {
-                            answer(true)
-                        }, declined: {
-                            answer(false)
-                        }
+                        accepted: { answer(true) },
+                        declined: { answer(false) }
                     )
                 }
             case let .upgradeOffer(offer, answer):
@@ -144,9 +136,7 @@ public class Glia {
                             answer(true, nil)
                             self?.callVisualizer.showVideoCallViewController()
                         },
-                        declined: {
-                            answer(false, nil)
-                        }
+                        declined: { answer(false, nil) }
                     )
                 }
             case let .videoStreamAdded(stream):
@@ -161,116 +151,12 @@ public class Glia {
         }
     }
 
-    /// Starts the engagement.
-    ///
-    /// - Parameters:
-    ///   - engagementKind: Engagement media type.
-    ///   - theme: A custom theme to use with the engagement.
-    ///   - visitorContext: Visitor context.
-    ///   - features: Set of features to be enabled in the SDK.
-    ///   - sceneProvider: Used to provide `UIWindowScene` to the framework. Defaults to the first active foreground scene.
-    ///
-    /// - throws:
-    ///   - `SalemoveSDK.ConfigurationError.invalidSite`
-    ///   - `SalemoveSDK.ConfigurationError.invalidEnvironment`
-    ///   - `SalemoveSDK.ConfigurationError.invalidAppToken`
-    ///   - `GliaError.engagementExists
-    ///   - `GliaError.sdkIsNotConfigured`
-    ///
-    /// - Important: Note, that `configure(with:queueID:visitorContext:)` must be called initially prior to this method,
-    /// because `GliaError.sdkIsNotConfigured` will occur otherwise.
-    ///
-    public func startEngagement(
-        engagementKind: EngagementKind,
-        theme: Theme = Theme(),
-        features: Features = .all,
-        sceneProvider: SceneProvider? = nil
-    ) throws {
-        guard engagement == .none else {
-            throw GliaError.engagementExists
-        }
-
-        guard let interactor = self.interactor else {
-            throw GliaError.sdkIsNotConfigured
-        }
-
-        let viewFactory = ViewFactory(
-            with: theme,
-            messageRenderer: messageRenderer,
-            environment: .init(
-                data: environment.data,
-                uuid: environment.uuid,
-                gcd: environment.gcd,
-                imageViewCache: environment.imageViewCache,
-                timerProviding: environment.timerProviding,
-                uiApplication: environment.uiApplication
-            )
-        )
-        startRootCoordinator(
-            with: interactor,
-            viewFactory: viewFactory,
-            sceneProvider: sceneProvider,
-            engagementKind: engagementKind,
-            features: features
-        )
-    }
-
-    /// Starts the engagement.
-    ///
-    /// - Parameters:
-    ///   - engagementKind: Engagement media type.
-    ///   - configuration: Engagement configuration.
-    ///   - queueID: Queue identifier.
-    ///   - visitorContext: Visitor context.
-    ///   - theme: A custom theme to use with the engagement.
-    ///   - sceneProvider: Used to provide `UIWindowScene` to the framework. Defaults to the first active foreground scene.
-    ///
-    /// - throws:
-    ///   - `SalemoveSDK.ConfigurationError.invalidSite`
-    ///   - `SalemoveSDK.ConfigurationError.invalidEnvironment`
-    ///   - `SalemoveSDK.ConfigurationError.invalidRegionEndpoint`
-    ///   - `SalemoveSDK.ConfigurationError.invalidSiteApiKey`
-    ///   - `SalemoveSDK.ConfigurationError.invalidAppToken`
-    ///   - `GliaError.engagementExists`
-    ///
-    public func start(
-        _ engagementKind: EngagementKind,
-        configuration: Configuration,
-        queueID: String,
-        visitorContext: VisitorContext?,
-        theme: Theme = Theme(),
-        features: Features = .all,
-        sceneProvider: SceneProvider? = nil
-    ) throws {
-        try configure(
-            with: configuration,
-            queueId: queueID,
-            visitorContext: visitorContext
-        ) { [weak self] in
-
-            self?.setChatMessageRenderer(messageRenderer: .webRenderer)
-
-            do {
-                try self?.startEngagement(
-                    engagementKind: engagementKind,
-                    theme: theme,
-                    features: features,
-                    sceneProvider: sceneProvider
-                )
-            } catch {
-                print("Engagement has not been started. Error='\(error)'.")
-            }
-        }
-    }
-
     /// Maximizes engagement view if ongoing engagment exists.
     /// Throws error if ongoing engagement not exist.
     /// Use this function for resuming engagement view If bubble is hidden programmatically and you need to
     /// present engagement view.
     public func resume() throws {
-        guard engagement != .none else {
-            throw GliaError.engagementNotExist
-        }
+        guard engagement != .none else { throw GliaError.engagementNotExist }
         rootCoordinator?.maximize()
     }
 
@@ -279,73 +165,8 @@ public class Glia {
     ///
     /// - Parameter messageRenderer: Custom message renderer.
     ///
-    public func setChatMessageRenderer(
-        messageRenderer: MessageRenderer
-    ) {
+    public func setChatMessageRenderer(messageRenderer: MessageRenderer) {
         self.messageRenderer = messageRenderer
-    }
-
-    private func startRootCoordinator(
-        with interactor: Interactor,
-        viewFactory: ViewFactory,
-        sceneProvider: SceneProvider?,
-        engagementKind: EngagementKind,
-        features: Features
-    ) {
-        rootCoordinator = self.environment.rootCoordinator(
-            interactor: interactor,
-            viewFactory: viewFactory,
-            sceneProvider: sceneProvider,
-            engagementKind: engagementKind,
-            screenShareHandler: environment.screenShareHandler,
-            features: features,
-            environment: .init(
-                fetchFile: environment.coreSdk.fetchFile,
-                sendSelectedOptionValue: environment.coreSdk.sendSelectedOptionValue,
-                uploadFileToEngagement: environment.coreSdk.uploadFileToEngagement,
-                audioSession: environment.audioSession,
-                uuid: environment.uuid,
-                fileManager: environment.fileManager,
-                data: environment.data,
-                date: environment.date,
-                gcd: environment.gcd,
-                localFileThumbnailQueue: environment.localFileThumbnailQueue,
-                uiImage: environment.uiImage,
-                createFileDownload: environment.createFileDownload,
-                loadChatMessagesFromHistory: environment.loadChatMessagesFromHistory,
-                timerProviding: environment.timerProviding,
-                fetchSiteConfigurations: environment.coreSdk.fetchSiteConfigurations,
-                getCurrentEngagement: environment.coreSdk.getCurrentEngagement,
-                submitSurveyAnswer: environment.coreSdk.submitSurveyAnswer,
-                uiApplication: environment.uiApplication,
-                fetchChatHistory: environment.coreSdk.fetchChatHistory,
-                listQueues: environment.coreSdk.listQueues,
-                sendSecureMessage: environment.coreSdk.sendSecureMessage,
-                createFileUploader: environment.createFileUploader,
-                createFileUploadListModel: environment.createFileUploadListModel,
-                uploadSecureFile: environment.coreSdk.uploadSecureFile
-            )
-        )
-        rootCoordinator?.delegate = { [weak self] event in
-            self?.handleCoordinatorEvent(event)
-        }
-        rootCoordinator?.start()
-    }
-
-    private func handleCoordinatorEvent(_ event: EngagementCoordinator.DelegateEvent) {
-        switch event {
-        case .started:
-            onEvent?(.started)
-        case .engagementChanged(let engagementKind):
-            onEvent?(.engagementChanged(engagementKind))
-        case .ended:
-            rootCoordinator = nil
-            onEvent?(.ended)
-        case .minimized:
-            onEvent?(.minimized)
-        case .maximized:
-            onEvent?(.maximized)
-        }
     }
 
     /// Clear visitor session
@@ -425,7 +246,6 @@ public class Glia {
 
     /// Ends active engagement if existing and closes Widgets SDK UI (includes bubble).
     public func endEngagement(_ completion: @escaping (Result<Void, Error>) -> Void) {
-
         defer {
             onEvent?(.ended)
             rootCoordinator = nil
@@ -437,11 +257,8 @@ public class Glia {
         }
 
         interactor?.endSession(
-            success: {
-                completion(.success(()))
-            }, failure: {
-                completion(.failure($0))
-            }
+            success: { completion(.success(())) },
+            failure: { completion(.failure($0)) }
         )
     }
